+++ Client/PandorasWallet/ServerAccess/ServerConnection.cs
@@ -523,7 +523,7 @@ namespace Pandora.Client.PandorasWallet.ServerAccess
 #endif
             string aTxWithNoFee = DirectCreateTransaction(aCurrencyTransaction);
 
-            int lFeePerKb = lCurrency.FeePerKb;
+            long lFeePerKb = lCurrency.FeePerKb;
             decimal lKBSize = ((decimal)aTxWithNoFee.Length / 2) / 1024;
 
             long lTxFee = Convert.ToInt64(lKBSize * lFeePerKb);
